+++ gatsby-config.js
@@ -12,7 +12,9 @@ module.exports = {
     description:
       'Pilvilinnojen maalauksen aika on ohi, me rakennamme sen, mitä toivot. Taitamme sinun visiosi – linssimme läpi.',
     twitterAuthor: '@visiosto_oy',
-    locales: ['fi', 'en'],
+    locales: ['fi', 'en-GB'],
+    localePaths: { fi: '', 'en-GB': 'en' },
+    simpleLocales: { fi: 'fi', 'en-GB': 'en' },
     defaultLocale: 'fi',
     defaultEmail: 'info@visiosto.fi',
     businessId: '3010084-6',